refactor(Test): move data into __data__ dir
