+++ tests/__data__/bip69-test-vectors.json
